Upgrade JDOM from 1.0 to 2.0.6.1
This upgrade fixes warnings in IntelliJ IDEA about the CVE-2021-33813 in the JDOM 1.0 dependency.

The <relativePath> tags are redundant and cause issues in IntelliJ IDEA.
It is also better to use JAVA_HOME instead of /usr/bin as javadoc may not always be symlinked to a Java 8 JDK causing compilation issues.

Signed-off-by: Wouter Born <github@maindrain.net>